build(about): combine ui/logic targets
And compile it statically.
