+++ lib/about/AboutDialog.qml
