+++ lib/about/git.cmake
