+++ lib/about/git.cmake.in
