+++ lib/about/version_gen.cpp.in
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "help/about/version.hpp"
+#include "about/version.hpp"
 #define GIT_SHA1 "@GIT_SHA1@"
 const char GIT_SHA1_tmp[] = GIT_SHA1;
 const char *const about::g_GIT_SHA1() { return GIT_SHA1_tmp; }
