+++ logic/help/CMakeLists.txt
@@ -1,2 +1,1 @@
-add_subdirectory(about)
 add_subdirectory(builtins)
