--- logic/help/about/dependencies.hpp
@@ -1,10 +1,0 @@
-#pragma once
-#include <QtCore>
-
-namespace about {
-struct Dependency {
-  QString name, url, licenseName, licenseSPDXID, licenseText;
-  bool devDependency;
-};
-QList<Dependency> dependencies();
-} // namespace about
