+++ ui/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_subdirectory(preferences)
 add_subdirectory(utils)
-add_subdirectory(about)
 add_subdirectory(text)
 add_subdirectory(help/)
 add_subdirectory(memory)
@@ -8,7 +7,7 @@ add_subdirectory(cpu)
 add_subdirectory(project)
 
 
-set(ui_targets "ui_utils;ui_about;text;memory;cpu;prefs;project;ui_help")
+set(ui_targets "ui_utils;text;memory;cpu;prefs;project;ui_help")
 
 
 if (APPLE)
